requirements.txt
decimal
tabulate==0.9.0
twilio==8.10.0
pytest==7.4.3
python-dotenv==1.0.0
